+++ platforms/Cross/plugins/SurfacePlugin/SurfacePlugin.h
@@ -6,17 +6,17 @@
 
 /* Plugins creating their own surfaces must register these using
    the following set of functions. The typedefs are for easier casts. */
-typedef int (*fn_getSurfaceFormat)(int surfaceHandle, int* width, int* height, int* depth, int* isMSB);
-typedef int (*fn_lockSurface)(int surfaceHandle, int *pitch, int x, int y, int w, int h);
-typedef int (*fn_unlockSurface)(int surfaceHandle, int x, int y, int w, int h);
-typedef int (*fn_showSurface)(int surfaceHandle, int x, int y, int w, int h);
+typedef long (*fn_getSurfaceFormat)(long surfaceHandle, long* width, long* height, long* depth, long* isMSB);
+typedef long (*fn_lockSurface)(long surfaceHandle, long *pitch, long x, long y, long w, long h);
+typedef long (*fn_unlockSurface)(long surfaceHandle, long x, long y, long w, long h);
+typedef long (*fn_showSurface)(long surfaceHandle, long x, long y, long w, long h);
 
 typedef struct sqSurfaceDispatch {
 	/* Version information. Must be provided by the client
 	   so the surface manager can check if certain operations
 	   are supported. */
-	int majorVersion;
-	int minorVersion;
+	long majorVersion;
+	long minorVersion;
 
 	/* Version 1.0 */
 	fn_getSurfaceFormat getSurfaceFormat;
@@ -27,14 +27,14 @@ typedef struct sqSurfaceDispatch {
 
 /* The functions for sqSurfaceDispatch are:
 
-	int getSurfaceFormat(int handle, int* width, int* height, int* depth, int* isMSB);
+	long getSurfaceFormat(long handle, long* width, long* height, long* depth, long* isMSB);
 		Return general information about the OS drawing surface.
 		Return true if successful, false otherwise.
 
 		The returned values describe the basic properties such as
 		width, height, depth and LSB vs. MSB pixels.
 
-	int lockSurface(int handle, int *pitch, int x, int y, int w, int h);
+	long lockSurface(long handle, long *pitch, long x, long y, long w, long h);
 		Lock the bits of the surface.
 		Return a pointer to the actual surface bits, or NULL on failure.
 		If successful, store the pitch of the surface (e.g., the bytes
@@ -65,14 +65,14 @@ typedef struct sqSurfaceDispatch {
 		be inside the source and dest boundingBox) but it is not aligned to word boundaries
 		yet. It is up to the support code to compute accurate alignment if necessary.
 
-	int unlockSurface(int handle, int x, int y, int w, int h);
+	long unlockSurface(long handle, long x, long y, long w, long h);
 		Unlock the bits of a (possibly modified) surface after BitBlt completed.
 		The return value is ignored.
 
 		The arguments provided specify the dirty region of the surface. If the
 		surface is unmodified all arguments are set to zero.
 
-	int showSurface(int handle, int x, int y, int w, int h);
+	long showSurface(long handle, long x, long y, long w, long h);
 		Display the contents of the surface on the actual screen.
 
 		If ioShowSurface() is called the surface in question represents
@@ -81,9 +81,9 @@ typedef struct sqSurfaceDispatch {
 	FXBlt uses a variant of the above functions which are exported from
 	the surface plugin:
 
-	int ioGetSurfaceFormat(int surfaceID, int* width, int* height, int* depth, int* isMSB);
-	int ioLockSurface(int surfaceID, int *pitch, int x, int y, int w, int h);
-	int ioUnlockSurface(int surfaceID, int x, int y, int w, int h);
+	long ioGetSurfaceFormat(long surfaceID, long* width, long* height, long* depth, long* isMSB);
+	long ioLockSurface(long surfaceID, long *pitch, long x, long y, long w, long h);
+	long ioUnlockSurface(long surfaceID, long x, long y, long w, long h);
 
 	These functions are looked up in the registered surfaces and invoked
 	as appropriate. The meaning of all values is exactly the same as for
@@ -93,7 +93,7 @@ typedef struct sqSurfaceDispatch {
 
 	Interpreter itself uses a separate entry point for updating the display
 
-	int ioShowSurface(int surfaceID, int x, int y, int w, int h);
+	long ioShowSurface(long surfaceID, long x, long y, long w, long h);
 
 	since the management of deferred updates is currently an intrinsic
 	property of the VM (which is bad - deferred updates should be a
@@ -104,17 +104,17 @@ typedef struct sqSurfaceDispatch {
 
 /* The following are the entry points for the surface manager:
 
-	int ioRegisterSurface(int surfaceHandle, sqSurfaceDispatch *fn, int *surfaceID);
+	long ioRegisterSurface(long surfaceHandle, sqSurfaceDispatch *fn, long *surfaceID);
 		Register a new surface with the given handle and
 		the set of surface functions. The new ID is returned
 		in surfaceID. Returns true if successful, false 
 		otherwise.
 
-	int ioUnregisterSurface(int surfaceID);
+	long ioUnregisterSurface(long surfaceID);
 		Unregister the surface with the given handle.
 		Returns true if successful, false otherwise.
 
-	int ioFindSurface(int surfaceID, sqSurfaceDispatch *fn, int *surfaceHandle);
+	long ioFindSurface(long surfaceID, sqSurfaceDispatch *fn, long *surfaceHandle);
 		Find the surface with the given ID, and, optionally,
 		the given set of surface functions. The registered handle
 		is returned in surfaceHandle. Return true if successful
@@ -124,8 +124,8 @@ typedef struct sqSurfaceDispatch {
 		interpreterProxy->ioLoadFunctionFrom("ioRegisterSurface","SurfacePlugin");
 	The typedefs below are for easier casts.
 */
-typedef int (*fn_ioRegisterSurface)(int surfaceHandle, sqSurfaceDispatch *fn, int *surfaceID);
-typedef int (*fn_ioUnregisterSurface)(int surfaceID);
-typedef int (*fn_ioFindSurface)(int surfaceID, sqSurfaceDispatch *fn, int *surfaceHandle);
+typedef long (*fn_ioRegisterSurface)(long surfaceHandle, sqSurfaceDispatch *fn, long *surfaceID);
+typedef long (*fn_ioUnregisterSurface)(long surfaceID);
+typedef long (*fn_ioFindSurface)(long surfaceID, sqSurfaceDispatch *fn, long *surfaceHandle);
 
 #endif /* __SQ_DRAW_SURFACE_H */
